Increase RLE offset and index types to uint16_t allowing for larger font data arrays (up to 65535 elements)

+++ Core/Inc/renderer.h
@@ -25,8 +25,8 @@
 /* typedefs ------------------------------------------------------------------*/
 typedef struct {
 	uint8_t ascii_code;         // The character this entry represents (e.g., 'A', 'a', '!')
-	uint8_t start_offset_rle;   // The starting index in the font's raw RLE data array
-    uint8_t end_offset_rle;     // The ending index in the font's raw RLE data array
+	uint16_t start_offset_rle;   // The starting index in the font's raw RLE data array
+    uint16_t end_offset_rle;     // The ending index in the font's raw RLE data array
     uint8_t col_width;          // The maximum width of the character in columns
 	uint8_t row_height;         // The height of the character in rows
 } Charachter_Font_t;
@@ -35,7 +35,7 @@ typedef struct {
     const char *name;
     const Charachter_Font_t *lookup_table;
     const size_t table_size; 
-    const uint8_t *rle_data;
+    const uint16_t *rle_data;
 } Font_t;
 
 /* Private Variables ---------------------------------------------------------*/
